requirements.txt
paho-mqtt
dotenv
PySimpleGUI
